+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes to Calva Backseat Driver
 ## [Unreleased]
 
 - [Update tool ids to qualify them as Clojure](https://github.com/BetterThanTomorrow/calva-backseat-driver/issues/34)
+- [Remove Human Intelligence tool](https://github.com/BetterThanTomorrow/calva-backseat-driver/issues/35)
 
 ## [v0.0.19] - 2025-09-11